refactor: split out about 400 lines from `lint.rs` into new modules
